[장바구니 협업 Step 2] 센트(김영우) 미션 제출합니다. (#125)
* refactor: 헤더 컴포넌트 수정

* refactor: 장바구니 페이지 반응형 디자인 추가

* feat: 모바일 환경 장바구니 페이지 하단 메뉴 추가

* feat: MSW API 추가

* feat: MSW API 완성

* refactor: MSW 주문 목록 불러오기 API 수정

* feat: 장바구니 페이지 주문하기 기능 추가

* fix: 적립된 적립금 바로 반영되지 않는 문제 수정 + 적립금 계산 오류 수정

* feat: 주문 목록 페이지 추가

* feat: 주문 내역 상세 페이지 추가

* fix: 선택된 상품이 없을 때 주문이 되던 오류 수정

* fix: 주문 완료 후 주문 목록 최신화 되지 않던 오류 수정

* fix: 주문 목록이 비어있을 때의 UI 추가

* refactor: 상품 삭제시 유저에게 한번 더 확인 받도록 수정

* refactor: Suspense, ErrorBoundary 적용

* fix: msw 에러 발생시키는 코드 해제

* refactor: atomFamily 사용해 서버 변경시 데이터 초기화 하도록 수정

* fix: 서버 변경시에 유저가 가진 적립금이 바뀌지 않는 오류 수정

* fix: 오타 수정

* style: console.log 삭제

* feat: 가장 최근에 선택한 서버 저장 기능 추가

* fix: 주문 상세 내용 요청 에러 수정

* fix: response header 정보 받아올 때 발생하는 에러 수정

* fix: github pages 새로고침 오류 수정

* style: 전반적인 스타일 수정

* fix: 주문 도중 발생하는 오류 대응

* refactor: 명세 변경에 따른 타입 수정

* fix: MSW 오류 수정

* style: favicon 등록

* refactor: 불필요한 Suspense 합치기

* refactor: 불필요한 null 병합 연산자 => default props로 변경

* refactor: 불필요한 null 병합 연산자 => props 기본값으로 변경

* refactor: <Fragment></Fragment> => <></>

* refactor: <Fragment></Fragment> => <></>

* refactor: useLocation 타입 지정

* refactor: 적절한 배열 메서드 사용하기

* refactor: atom effects활용해 localStorage에 atom상태 저장

* refactor: fetch의 body를 만드는 부분 함수 분리

* style: useState<number>(0) => useState(0)

* style: if문 내의 return 부분 수정

+++ src/components/mainPage/productCard/AddCartButton.tsx
@@ -1,18 +1,20 @@
 import { useState } from 'react';
-import { ShoppingCartIcon } from '../../../assets/ShoppingCartIcon';
 import { useCartRecoil } from '../../../hooks/recoil/useCartRecoil';
 import { Counter } from '../../../layout/counter/Counter';
 import { useCartFetch } from '../../../hooks/fetch/useCartFetch';
-import Loading from '../../common/Loading';
+import { Loading } from '../../common/Loading';
 import { useRecoilValue } from 'recoil';
 import { cartItemsState } from '../../../recoil/atoms/cartAtom';
+import styled from 'styled-components';
+import { APIAtom } from '../../../recoil/atoms/serverAtom';
 
 interface AddCartButtonProps {
   productId: number;
 }
 
 export const AddCartButton = ({ productId }: AddCartButtonProps) => {
-  const cartItems = useRecoilValue(cartItemsState);
+  const apiEndPoint = useRecoilValue(APIAtom);
+  const cartItems = useRecoilValue(cartItemsState(apiEndPoint));
   const initialQuantity =
     cartItems.find((cartItem) => cartItem.product.id === productId)?.quantity ??
     1;
@@ -45,6 +47,11 @@ export const AddCartButton = ({ productId }: AddCartButtonProps) => {
   const deleteCartItem = () => {
     if (cartId === undefined) return;
 
+    // eslint-disable-next-line no-restricted-globals
+    const isUserWantDelete = confirm('상품을 장바구니에서 삭제하시겠습니까?');
+
+    if (!isUserWantDelete) return;
+
     deleteRecoilCartById(cartId);
     deleteCartItemById(cartId);
   };
@@ -72,10 +79,36 @@ export const AddCartButton = ({ productId }: AddCartButtonProps) => {
           onQuantityChange={handleChangeQuantity}
         />
       ) : isLoading ? (
-        <Loading />
+        <Style.LoadingBackground>
+          <Loading width={20} height={20} />
+        </Style.LoadingBackground>
       ) : (
-        <ShoppingCartIcon handleClick={handleClickShoppingCartIcon} />
+        <Style.ShoppingCartImage
+          src={`${process.env.PUBLIC_URL}/plusIcon.svg`}
+          alt="장바구니 추가 버튼"
+          onClick={handleClickShoppingCartIcon}
+        />
       )}
     </>
   );
+};
+
+const Style = {
+  ShoppingCartImage: styled.img`
+    width: 35px;
+    height: 35px;
+
+    cursor: pointer;
+  `,
+  LoadingBackground: styled.div`
+    width: 35px;
+    height: 35px;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    border-radius: 50%;
+    background-color: white;
+  `,
 };
